Refactor winres functions and source code
